Refactor models into a separate directory and update references
- Moved models from flowapp/models.py into a dedicated directory
- Updated flowapp/output.py to handle rule_type as an enum value
- Fixed a return value issue in flowapp/utils.py's parse_api_time function
- Added a debug print statement in flowapp/views/rules.py for user actions UI
